Update apache worker proc config to match optimus.
On srobo-devel a long while back, I suggested that we didn't need a vast
number of spare httpd processes hanging around consuming memory, and
that we could decrease the number of spares required. This was met with
approval and deployed to optimus; it's survived kickstart and a
competition since then, so I'm pretty certain that this config works.

Specifically, this patch decreases the number of spare server
proceesses maintained; decreases how many are allowed to hang around
being spare after requests complete; and limits the maximum number of
server processes permitted.