Configure svn installation.
Files belong to svn, group svn, and apache is placed in that group.

File permissions are borked. I tried configuring so that everything got
group sticky bit, but that resulted in the execute bit being unset for
some directories, and /all the files/ being setgid too. Which would be
silly. So, it's left as 644, which is fine because SVN is readonly.