Install git repo scripts.
I originally wrote some exects to fix the permissions of this whenever
it was updated; however it makes much more sense to make the scripts
checkout managed only by puppet (and not edited in place by operators),
which would mean that no permissions fail had to occur.

This also means that we should be tracking a normal user git repository
somewhere that contains these scripts. Which has yet to happen.