Actually, install voting stuff in a users homedir
This is because the apache goo require to make suexec work turns out to
be difficult to configure on a per-file or per-dir basis. It can only
(apparently) work on vhosts or on user dirs. So to keep the owner of the
voting files seperate from the rest of the website, it _has_ to go in a
home directory.