Actually, use 'apache' instead of www group.
Two things: first, apache has to be the member of whatever group owns
all the web files, because otherwise it can't read secret config files
(i.e. phpbb's config.php). This is ok because it still can't write to
things with group permissions.

Second, while we could use a 'www' group anyway, we'd have to configure
apache (the user) to be part of it, and because puppet is rubbish it
can't do that without managing the apache user itself. Bleaguh.