Allow wwwcontent to log in.
So, we have a problem. I've been configuring the owner/group attributes
of vcsrepos so that they belong to the desired user/group.
Unfortunately it turns out that vcsrepo is not particularly clever on
how it applies this; it just chown -R's everything. Which is bad.

This means that when puppet reapplies its configuration, it resets the
owner of a whole host of ide data files that are supposed to be owned by
apache so that it can fiddle with them. This is bad. This also happens
to piwik too.

The solution is to make all vcsrepo processes run as the relevant user
so that things are owned by the correct user; and we don't configure the
'owner' field, that way no bulk rewriting occurs.