Make /var/www belong to wwwcontent.
It turns out that the vcsrepo module deletes the directory it's going to
clone into before cloning it. Which makes my attempts to ensure that it
can clone the desired repository as the wwwcontent user pointless.
Uuuugh.

Instead fix this by making /var/www belong to wwwcontent, and clone
everything else into it. Sigh.