Install the gerrit service script.
Recommendation is to symlink it, but I worry that that'll lead to
something owned by gerrit being run by root, which is bad.

There's the potential for installing this as systemd in the future,
but I don't know how that works, and *do* know how SYSV init scripts
work, so stick with that for now.