Install the gerrit/java mysql connector.
A vastly horrible operation: download it from where gerrit downloads
it from, and verify its checksum too. Why gerrit doesn't install this
by default I don't know.

A significant worry is that we'll forget to update this as we update
gerrit in the future. However it appears there are going to be a hell
of a lot more problems with gerrit, seeing how between releases they
make schema changes, and you have to re-run the initialise program
to update the schema. Which is survivable, but unpleasent.