Add a script for "fixing" a gerrit config file.
After creating a gerrit installation, one has to put a mysql database
into the corresponding database. This is fine and is supplied in the
server dummy secrets.

Unfortunately it seems Gerrit replicates some of the data in it's
config database, dumping a randomly generating guid into it's
all-projects git repo. If the guid in the database and in the repo
don't match, then the administrator group gets permanently buggered.
So, this script patches the config repo.