Install the backup scripts.
Note that it's configured to not track origin/master, and thus to
not auto update. This is a feature: under the scheme currently being
discussed, scripts in backup.git might end up being run as root by
a remote cron, in which case auto-updating to something that someones
pushed is an easy route to root.