Allow backup user to run backups without a tty.
I added a check to the backup scripts so that it wouldn't dump a tar
onto stdout if it's a terminal. However our sudo configuration
requires that sudo be run with a terminal, which is mutually
exclusive.

So, allow just the backup user to run sudo without a tty, making this
fully scriptable.