Prevent puppet from resetting trac db charset.
What happens here, is that puppet inspects the trac db, sees that part
of its character set is set to utf8_bin, and resets it to utf8. Then
the sql in /tmp/trac.init runs again, resetting the charset to utf8
and utf8_bin. This is pointless, and happens every time puppet is
applied.

So instead, set a charset value that leads to puppet not resetting the
charset every time it inspects the database. The DB has exactly the
same state, just now there are fewer notices when apply a config.