Remove trac charset setting completely.
We can't have puppet trying to make it utf8_bin unfortuantely, or it
tries to set both the character set and COLLATE to utf8_bin, which
doesn't work.

Basically I don't fully understand what's being done here, so instead
just stick with a configuration that works, and swallow the additional
messages that pop up whenever puppet is reapplied.