Start using www_hostname for http config.
So: I scattered a bunch of config variables for the webserver config,
so that it had a bunch of urls and hostnames encoded into it at various
points. The one thing that's common to all of them is the webserver
hostname, and in fact we can extrapolate all other pieces of data from
that.

So, start working towards that situation.