svn initialisation: Improve error catching/cleanup
If the svn initialisation failed because the source dump file was not
found, the initialisation would proceed as if it had, and
/usr/local/var/sr/svn_installed would get touched.

This patch catches errors in the gzipping process, and avoids that
file being created etc.  It also adds cleaning up of the temporary
gunzip'ed version of the svn dump in the event of an error.