Fix #1407: Add a local srweb config.
This allows easier deployment of srweb, as it removes the need to
modify the local config.inc.php, which would otherwise need to be
done after each puppeted update.

This relies on srweb patch f185ae68d43764da34 (gerrit:#160), which
adds support for local config files in this manner. I've also added
the requisite variable to common.csv in dummy-secrets ebacf60
(gerrit:#164) though I obviously can't add this to the real secrets.