Ensure that /var/www/html is owned by wwwcontent.
I thought that as /var/www was owned by wwwcontent, when srweb was
cloned into /var/www/html it'd remove that directory and replace it
with another one owned by wwwcontent. However what appears to have
happened is that vcsrepo wipes the dir (as root?) and then clones
into it as wwwcontent?

Either way, something bad has happened, that results in /var/www/html
being owned by root and wwwcontent not being able to write to it. Which
means that srweb changes to /var/www/html that attempt to delete or
create files break, as wwwcontent doesn't have permission.

This isn't a state I think is valid to be created, but there it is.
Fix by making everything more explicit.