Enable HTTP keep-alives.
This prevents apache from sending a 'Connection: close' header with
every request. Previously this meant that every single request had to do
a full SSL negotiation, which was exceedingly bad for everyone and
everything.

The two configurable values are number of requests and time to keep
the connection open for. I picked something arbitarily large for the
number of requests, and some reasonable portion of time for the
connection timeout.

The only constraint I'm aware of is the IDE polling frequency of 7s, and
I've followed Professor Brown's principal of 'Pick a reasonable then
double it'.