Bump up httpd server limit.
The Bristol Tech day is tomorrow, and we've already seen that the past
bad configuration led to the website being unaccessible, which was fixed
by increasing the number of httpd servers available.

I won't be at the tech day; we know this change fixes the outage and
doesn't kill the website (at least under tech day load); so
pre-emptively apply it. We'll be able to see from memory usage graphs
whether the previous keep-alive and WSGI fixes have had the desired
effect.