Force IPV4 connection to Gerrit.
This has suddenly stopped working after upgrading the kernel on badger
from 3.7 to 3.8: apache attempts to connect over v6, gerrit isn't
listening on it. This worked in the past; nothing has changed but the
kernel update; I don't know how to tell gerrit to try harder, so fall
back to this hack instead.