Fix #1540: Add support for using pipebot's config setup.
This replaces the previous assumption that the default config was
correct, and thus resolves the contention for names on freenode.
While all the dev pipebots will still clash, this is far less of an
issue than dev pipebots clashing with the live one.

Note that this requires a couple of new lines in the common.csv,
which are present in the dummy-secrets (see gerrit:558 & 559), but
will need adding to the live ones.

Change-Id: I153a6a57245a30aa4f791004acb84cba40aeaa6a