Fix #1178: Allow mentors RO access to team git repos.
We trust Blueshirts not to share this access with teams,
and it potentially allows for easier mentoring if they can
all access all the teams' source code. This introduces read-only
access via the 'mentors' ldap group and a separate url.

Change how the template and the helper script work, such that the
LDAP anon password, which is now needed twice, is injected into the
config direclty by puppet, rather than going via the pythong script.

Change-Id: I78e99c1065ca76ea7e6fedbf0e2a61df619a900e