Ensure that IDE internals can't be accessed.
Prevent access to raw repos, settings files (which contains
team-status data) & notifications. Also prevent listing zip
files (since these also contain team code) & copy in the
contents of the .htaccess for the zips folder that's been
in the IDE sources for a while now, but doesn't seem to be
present live.

Change-Id: I7af2d072b9276f385be082e0cd9afedb874ca2fd