Set TEAM_STATUS_DIR in srweb local config
As described in #2156, we are arranging for srweb.git to by default
contain some dummy team data that'll allow the team-status pages to be
tested without an IDE set-up.  To do this, the default location of the
team-status information needs to change.  We don't want to change the
location of our live system's information, so before we make this
change, this patch moves the current default for this setting into the
production srweb's local config.