Update gerrit to gerrit 2.7
I'm more or less shooting from the hip here: the main webserver has been
updating to 2.7, and I believe the changes I've made here are
consistent with what happened there. This is not an automated update
though: a schema update must still be applied.

One piece of errata is that gerrit now requires the java mysql connector
version 5.1.21; I'll ticket the update for this.