Swap out ntpd for chrony.
This is apparently desired as it's easier to configure and makes
interacting with the other competition machines easier.

Tested by manually removing the other timing sync things in my badger
clone (including ntpd), setting the date to the start of this month,
and then applying this patch. The desired effect of the date being
corrected was observed, so this appears to work in the simple case.

Change-Id: Ia557b8bcc5e7cd5251aa0d85a9fae8b43151b5f3