Remove (most of) the constant list of dbs to back up.
It had already got out of date with respect to the phpBB database.
By building it from other variables, we avoid that occuring again.

It would be nice to have variables for trac and gerrit too, which
this patch creates in the right place, but due to a WONTFIX bug in
puppet (https://tickets.puppetlabs.com/browse/PUP-1397), hyphens in
names cause various issues.

Change-Id: Ie026b8e706ef35a6258e5a66d929169fb39ae9a2