Fix #2489: Update python docs fetching and use wget
This changes the url we use to download the python docs from so that
it's the one currently published, rather than being the wrong side
of a bunch of redirects.

Also change from using curl to using wget since wget deals with the
redirection transparently by default, whereas even after reading the
man page and several minutes of trying I couldn't make curl do so.

Change-Id: Ic7afbe50765a450456f55b7a9012b988f3204f49