Don't force the creation of git working copies.
It turns out that the 'correct' behaviour of this value is to delete
and then re-create the clones each time, which clearly isn't what
we want. Thankfully that behaviour isn't present in the version of
the vcsrepo module we're using at the moment, but we're going to
end up switching at some point.

This shows up some dependancy problems with /srv/srcomp-http and
/var/www/html, where things were installed into the location of git
checkout's before the vcsrepo was installed. Previously vcsrepo just
rm'd that directory first (which explains some past experiences...),
but now it refuses to install instead (which is good).

Resolve this by deleting /var/www/html, having srweb depend on
/var/www's permission change, and paraphanalia installed inside
srweb and srcomp-http depending on the outer vcsrepo.

Change-Id: Ie8be0881db04765759125c0724ee6b88306254c0