Don't force the creation of git working copies.
It turns out that the 'correct' behaviour of this value is to delete
and then re-create the clones each time, which clearly isn't what
we want. Thankfully that behaviour isn't present in the version of
the vcsrepo module we're using at the moment, but we're going to
end up switching at some point.

Change-Id: Ie8be0881db04765759125c0724ee6b88306254c0