Fix #1446: Get trac's config into puppet.
This has the trac manifest create the trac environment on the fly,
rather than 'restoring' the whole thing from backup. We still copy
the attachments folder from the backup as it is content and not
configuration.

This commit includes a copy of the config files from badger, converted
to have erb placeholders instead of the actual secrets. It also
includes the site template and trac logo, neither of which are
particularly secret.

Change-Id: I289a640cc2efcd7ca9d9efae2026e6c7c59ef876