Don't maintain trac attachments in puppet
Trac allows for deleting attachments and reuploading them, changing
their contents. Puppet will reset that if it sees it, so instead
manually copy attachments to the right location on installation, and
maintain their permissions with puppet.

As with other data-installation stuff in our puppet configuration, touch
a file in /usr/local/var to guard puppet from reinstalling stuff on the
top of existing data.

Change-Id: If9ff6dca4346e7a1916ea0facea363e245b7133a