Account for systemd in PAM logins
Invoking this module on session opening ensures that users who log in
get their own cgroup. Without this, all user processes that are
descendants of an ssh login (inc tmux and screen) are wiped when sshd
restarts. This had led to an intense amount of grief when I test things
on vagrant in the past, which I wasn't able to debug until I stumbled
across someone having the same problem.

Of course, only systemd would decide that if their model of operation
doesn't apply to a particular use case, the use case has to change,
rather than making the model configurable.

Change-Id: I1cf01f92b7350469e18da541f5279be09891c7bb