Fix team-status-backup-restore further
It was switched to running as apache to avoid overwriting files as root,
in a non-root-owned directory. However, that then means that it can't
set the 'installed' flag in /usr/local/var/sr. And we don't want
unprivileged processes writing that, seeing how they can force a reset
of installed state.

So instead run the restore script as root; but have it drop privs to
apache when it actually performs the copying. Uuuggh.

Change-Id: I8e1c4d1276d7f9575116bbb48b043a8977888b50