Trac needs both authentication and authorization
It turns out that in apache 2.4, 'Require' is implicitly inside a
RequireAny block, meaning that if any 'Require' directive is matched,
access is permitted. That means valid users can skip the LDAP group
membership test for trac.

Fix this by enclosing the 'require' directives in a RequireAll block,
which requires all it's internal directives to succeed.

(NB: The apache documentation states that, actually, RequireAll means
"one require-directive must succeed, and the others must not fail". I
don'really understand how a neutral authn/authz response can be
overlooked in the context of "All").

Change-Id: Icb24cfd88836af22490be77619d7054a920fadb6