Aimlessly rotate gerrit config to appease linux
After the webserver got v3.17, all of a sudden the old way of binding to
addresses stopped working, for reasons I do not understand.

Gerrit binds to all addresses now. this means we are relying on the
firewall to block incoming http to it; in particular the default fedora
v6 firewall that we don't explicitly configure.

Change-Id: I3233f898038f482d1227d5e9a918763eca912957