Inject a Willie into Saffron
This adds a Willie IRC bot instance onto the main server and
configures it to act as a replacement srbot. Following the pattern
established by pipebot, the service is neither installed nor enabled
in devmode, so this needs modifying for testing (as does the bot name).

This also has the srbot-plugins repo auto-deploy, though it doesn't
currently contain any python modules yet.

Change-Id: Ia5632712b191d4746eca170bda69c0c25adcf01b