Fix installation of the All-Projects.git
Having changed its intallation to error properly it becomes clear
that that installation wasn't working, and it turns out that this
was due to the gerrit user (correctly) not being able to actually
access the all_projs.tgz file in the secrets. Un-tar the file as
root then chmod the directory in order to fix this.

This will ensure that dev clones of the server have a working
gerrit install without additional permission munging.

Change-Id: I62123c22e0b9e110a0dff28d998aaad07db61ecf