Remove the IDE Repo 'conf' script
This appears to be rather out of date and would most likely break
things if anyone actually ran it. Specifically the repos are now
created with the proper shared/denyNonFastForwards values anyway
but the group owning them is 'apache'. AFAIK 'www' doesn't exist
as a group on saffron. I'm admitedly unsure what the other permission
munging this does is meant to achieve, but since things are working
without them I think we can safely remove this file.

Change-Id: I5f1ad12fdac31671232546d4ce8e6357d638de74