Also rquire gcc for the YAML acceleration
It turns out that python-devel doesn't depend on there being a
compiler around, so we've ended up with only the pure-python YAML
loader around despite having tried to use the C extension.

While installing GCC now won't fix that, it will ensure that any
builds after now will work as designed, and make it possible to
reinstall the components in the virtualenv which rely on there
being a compiler to build their C extensions.

Change-Id: Iee2cc49efd3f36367ca246657c9ae93af2a1e002