Import configuration for an IDE specific apache
These config files create another apache server process, but crucially
that does not listen on port 80, or store its PID file in the same
place. It's then configured to load IDE specific configuration, and have
it's own dedicated queue of workers for doing IDE like things.

Deploying this to be a service is slightly more involved.

Change-Id: Ifdec95170ae665218f8ba8bb37594e2a44ba733f