Transition to serving the website via nginx
This turns on nginx, listening on port 443, which promptly reverse
proxies data to a) the main website on port 8000 over http, and b) a
dedicated IDE apache worker on port 8001. The main apache instance also
serves redirects via port 80.

This should load balance the IDE a little: worker threads in apache will
be reserved for IDE purposes. It will still be possible to access the
IDE via the main apache instance, however only if you access it from the
webserver itself.

Ports 8000 and 8001 will remain firewalled off. SSL is turned off on
those apache instances to avoid having to have two SSL engines running,
with a redundant SSLing of the reverse-proxy connection.

Change-Id: Id06f0b3f1cfddf9619e95cd34bfbfba6566e3c03