Start importing common.csv info into hiera
That is; rather than using a CSV file to store configuration data,
instead store it in a hiera yaml file. This is because later versions
of puppet do not support the 'extlookup' function.

This means that some dummy configuration is in puppet, and some is in
the dummy-secrets repo. This is too bad: there's no elegant way of
solving the fact we have random binary blobs to distribute, so this is a
reasonable compromise.

Exactly how the production-specific configuration is to be installed,
and how to protect it, is an open problem.