Fix a namespace mangle in ide's httpd config file
The specific issue is that the web root directory variable is
web_root_dir in the httpd module, but in the ide httpd module it's
called root_dir. Previously, puppets broken namespace situation meant
that this could be resolved across class boundaries (or something). Now
the 'web_root_dir' variable expands to nothing in the IDE's ssl.conf,
and as a result things break.