Update the forums and move to extensions
This moves us from our locally modified 3.0.x phpBB to a plain
3.1.6 checkout plus extensions. The 3.1.x model for handling
change is now via extensions and this should make future upgrades
somewhat easier.

There are two extensions which we're using here, both are currently
only hosted on my github though could easily (and probably should)
move to being SR-hosted. The first adds the pipebot integration,
the latter adds everything else.

This also introduces the 'archive' module which makes the download
and extraction of external archives somewhat cleaner. This is used
to get the theme we use.

Change-Id: Iae3a0304993ed3d0f6158f91662992c907b4fee3