Revert "Disable fritter cron emails"
This reverts commit e28afd50a3efa2c20dd7b495bb79b0e513da3cd5.

This was an issue that was going to crop up in a parallel branch,
whereby the moment sendmail was installed, one million fritter emails
would charge out onto the internet. However, now that postfix is being
used there, this isn't going to happen.

(Because /usr/sbin/sendmail is for esmp, and postfix is using a
different binary).