Refine authz config for RT
Everyone is matching the 'localhost' permission and not getting a
REMOTE_USER environment. This is because we reverse proxy via nginx,
through localhost. I've tried to block this through requiring that
X-Forwarded-For isn't set, but with no success. Yay reverse proxies :/

Refine this by overriding the authz configuration in the location for
the unauthenticated RT REST API. It needs to be able to call it from
localhost: allow that, block everyone else.

Change-Id: I605c392200f3510d345d337cff680b4fc953b5fd