Introduce a new 'voting' group to make voting work
Following the move to F22, the voting system doesn't work due to
apache not being able to suexec to the voting user. This turned
out to be due to its group being gid 100 which is apparently
restricted. Avoid this by giving the voting user its own group
with a (fixed) higher id.

Change-Id: Id4f2f6d3610339c7f4a2980fea217cbe80f75bfe