Fix #3058: Fix commentlink overmatch on hashes
Gerrit appears to run the commentlink matches on the raw html rather
than the rendered text. As a result of this matches which look for
special HTML characters (such as hashes) cause weirdness in the
resulting rendering. I've reported this upstream [1], but for now
we can work around this by excluding tickets preceded by ampersands
which seems fairly unlikely anyway.

[1] https://code.google.com/p/gerrit/issues/detail?id=3629

Change-Id: I008770589a131125c9aef44d47b898ae41b7bd3b