Explicitly ignore newlines in trac
This config option is a little complex in how it behaves by
default as it has different default values based on the history
of the trac instance. Since the live site pre-dates it, we get
the 'no' behaviour, but any newly created instances (including
any created via this puppet setup) get the 'yes' behaviour.

The value we're used to is 'no', so explicitly set that.

Change-Id: I3beb78c0a5e8995a812ffa66bc0886eaf42a244a