Support HSTS preloading
This is deliberately a separate commit as enabling preload has
somewhat long-term consequences. With this enabled we can submit
our domain to various browser preload lists. This ensures that
all visitors (even in clean browsers) will primarily connect
using HTTPS, which offers a marginal security improvement.

It should be noted that while addition to the preload lists is
fairly easy, removal is more complicated. As we're not likely
to want to be removed any time soon I don't believe this is
an issue. For further discussion, see the thread 'HSTS preload
and subdomains' on srobo-devel:
http://thread.gmane.org/gmane.science.robotics.srobo.devel/1290.

Change-Id: I68d97939cd41d66aebabbc9f6e83a85542385155