Disable subversion via config
This more easily allows for the disabling of chunks of the Apache
HTTPD configs in step with the disabling the underlying manifest.

For the moment we disable this everywhere other than saffron as
it doesn't work on recent Fedora versions. Saffron is still
running an old Fedora, though hasn't actually been updateed to use
this config yet.