Upgrade to TLS without redirection to root domain
We use puppet only for pet machines hosting services, rather than
as the main website host now, so we don't need to redirect people
to the main website from their insecure access.

We may still want to redirect users to the canonical domain once
they're on a secure connection, however that's a separate concern.