Add support for certificates from letscenrypt
We use the apache server instance (via webroot) to complete the
domain challenge because it's already on port 80 (and for the
moment needs to stay there for the SVN support) even though the
nginx server is the one which eventually uses the certs. This is
definitely a bit of a mess, though hopefully we can make most of
the apache intermediate go away by moving mostly to nginx sometime
soon.